Defined a 'next' method.  Test failing.
We have defined a method on the Counter class which does nothing.

The 'rb_define_method' associates the Ruby string (name of method)
with the C function.  As C is not an object-oriented language, all
the functions (in a Ruby extension) take the object on which they
act as the first parameter.  The fourth parameter of '0' indicates
that the (Ruby) method takes no parameters.  Thus the C function
has one parameter 'self' (the object associated with the call).
The type of the values coming in from and returning to the Ruby
environment are 'VALUE'.  More on this later.

All (Ruby) methods must return a value (unlike C functions): if
nothing else, a nil.  Produced in C with 'Qnil'.