Added an initialize method.
We need an initialize method in the C code to have somewhere we
can allocate the memory which will hold the class's state.

'initialize' is a method like any other, and needs to return a
VALUE.  Traditionally (ie in the Pickaxe) you return 'self', but
'Qnil' seems to work as well.