Allocate a structure to hold the state.
We define a structure (currently empty) to hold the state of the
Counter object.  I have called the C function which allocates the
memory for this 'allocate_structure'.  It is tied to the Ruby
infrastructure with the 'rb_define_alloc_func' call.  Having
allocated the structure, it is wrapped so Ruby can handle it with
the 'Data_Wrap_Struct' call.  This takes four parameters.  The
class_name is passed in because we may have been subclassed.  A
mark routine would be required if our data held references to Ruby
objects.  (We won't get into that.)