Getting messy now.  Tests failing.
Basically, this shows that I didn't need to introduce 'initialize'
yet.  I can retrieve 'count' within the 'next' function.

Note that I am using the C functions 'calloc' and 'free', rather than
the Ruby-garbage-collector-aware ones.

I also need to explain the naming of 'cp'.  (Which rather indicates
that it is a bad name.)