Can now specify starting value for the count.
The initialize method is no different to any other method, it just
gets called by 'new' after the memory is allocated.

I use this as an opportunity to demonstrate how to pass optional
parameters to a method implemented in C.  The argument count parameter
of '-1' specifies that all parameters are optional.  ('-2' would
specify that there was 1 mandatory parameter followed by optional
ones, and so on.)

The signature of the target C function changes to receive the count of
arguments and an array of their VALUEs.  The 'rb_scan_args' function
takes these parameters with a specification string, checks for errors,
and puts the values into specified variables.  The specification "01"
specifies 0 mandatory and 1 optional parameter  Errors are reported
by exceptions with helpful messages.  Optional parameters which are
not provided result in a nil value.  The macro NIL_P tests for nil.